feat: 實作監控頁面的噴嘴移動與 Z 軸偏移控制功能
Why:
- 實作噴嘴 X/Y/Z 軸移動、XY 歸零、Z 軸偏移調整等控制功能，使 UI 控件輸入的數值能夠正確傳送到機器執行
- 將 `Bed Movement` 相關變數/函式/枚舉項目重命名為 `NozzleOffset`，因為實際功能是調整座標系統的 Z 軸偏移而非列印平台移動
- 確保 G-code 命令格式：WebSocket 傳送時不使用 URL 編碼
- 修正X/Y/Z 軸移動switch case 0.1mm回傳的數值

How:
- 實作 `SetPhrozenCommand_nozzle_movement`：支援 X/Y/Z 軸方向移動（透過 `direction` 參數指定 "x", "y", "z"）與 XY 軸歸零（"home_xy"）
- 實作 `SetPhrozenCommand_nozzle_offset`：支援 Z 軸偏移調整
- 修正 `zoffset` 函數：參數類型從 `int` 改為 `float`，新增 `MOVE=1` 參數，並確保不含URL 編碼的空格`%20`符號
- 修正 `get_selected_nozzle_movement_range`：將 `Range_01_MM` 返回值從 `0.01f` 修正為 `0.1f`
- 重新命名：將 `PhrozenPrintBedMoveRange` 重命名為 `PhrozenPrintNozzleOffsetRange`，相關變數/函式/枚舉項目同步重命名為 `NozzleOffset` 相關命名，並新增 `Nozzle_Home_XY` 枚舉項目
  (ex: PhrozenPrintBedMoveRange, m_kBedMovementRangeButtons, get_selected_bed_movement_range,  on_bed_movement_range_mouse_left_down, PhrozenMovement::Bed_Z_Positive/Negative)
- 在 `on_manual_movement_changed` 中實作所有移動命令的調用，新增 macOS 特定代碼獲取 `PhrozenMachineObject` 對象

+++ src/slic3r/GUI/PhrozenGUI/PhrozenMonitorController.hpp
@@ -848,7 +848,7 @@ struct WebCamImageDataThreadHandler
     
     bool homeZ_http();
     
-    CURLcode zoffset(int value);
+    CURLcode zoffset(float value);
     
     CURLcode SetExtruderTemperature(int value);
